+++ app/services/forest_liana/intercom_conversations_getter.rb
@@ -31,6 +31,7 @@ module ForestLiana
       rescue Intercom::ResourceNotFound
         @records = []
       rescue Intercom::UnexpectedError => exception
+        FOREST_REPORTER.report exception
         FOREST_LOGGER.error "Cannot retrieve the Intercom conversations: #{exception.message}"
         @records = []
       end