+++ app/services/forest_liana/token.rb
@@ -20,6 +20,7 @@ module ForestLiana
         last_name: user['last_name'],
         team: user['teams'][0],
         role: user['role'],
+        tags: user['tags'],
         rendering_id: rendering_id,
         exp: expiration_in_seconds()
       }, ForestLiana.auth_secret, 'HS256')