+++ lib/forest_liana.rb
@@ -28,6 +28,7 @@ module ForestLiana
   mattr_accessor :names_overriden
   mattr_accessor :meta
   mattr_accessor :logger
+  mattr_accessor :reporter
   # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
   mattr_accessor :names_old_overriden
 
@@ -40,6 +41,7 @@ module ForestLiana
   self.names_overriden = {}
   self.meta = {}
   self.logger = nil
+  self.reporter = nil
 
   @config_dir = 'lib/forest_liana/**/*.rb'