+++ lib/forest_liana/bootstrapper.rb
@@ -57,10 +57,12 @@ module ForestLiana
       @collections_sent.each do |collection|
         collection['actions'].each do |action|
           c = get_collection(collection['name'])
-          a = get_action(c, action['name'])
-          load = !a.hooks.nil? && a.hooks.key?(:load) && a.hooks[:load].is_a?(Proc)
-          change = !a.hooks.nil? && a.hooks.key?(:change) && a.hooks[:change].is_a?(Hash) ? a.hooks[:change].keys : []
-          action['hooks'] = {'load' => load, 'change' => change}
+          unless c.nil?
+            a = get_action(c, action['name'])
+            load = !a.hooks.nil? && a.hooks.key?(:load) && a.hooks[:load].is_a?(Proc)
+            change = !a.hooks.nil? && a.hooks.key?(:change) && a.hooks[:change].is_a?(Hash) ? a.hooks[:change].keys : []
+            action['hooks'] = {'load' => load, 'change' => change}
+          end
         end
       end
     end
@@ -82,7 +84,8 @@ module ForestLiana
             @collections_sent = content['collections']
             @meta_sent = content['meta']
             generate_action_hooks
-          rescue JSON::JSONError
+          rescue JSON::JSONError => error
+            FOREST_REPORTER.report error
             FOREST_LOGGER.error "The content of .forestadmin-schema.json file is not a correct JSON."
             FOREST_LOGGER.error "The schema cannot be synchronized with Forest Admin servers."
           end
@@ -122,6 +125,7 @@ module ForestLiana
           end
         end
       rescue => exception
+        FOREST_REPORTER.report exception
         FOREST_LOGGER.error "Cannot fetch properly model #{model.name}:\n" \
           "#{exception}"
       end