+++ spec/services/forest_liana/permissions_checker_acl_disabled_spec.rb
@@ -481,6 +481,14 @@ module ForestLiana
               end
             end
 
+            context 'when user has no segments and param segmentQuery is there' do
+              let(:segmentQuery) { 'SELECT * FROM products;' }
+              let(:collection_list_parameters) { { :user_id => "1", :segmentQuery => segmentQuery } }
+              it 'should be authorized' do
+                expect(subject.is_authorized?).to be false
+              end
+            end
+
             context 'when segments are defined' do
               let(:segments_permissions) { ['SELECT * FROM products;', 'SELECT * FROM sellers;'] }
               let(:collection_list_parameters) { { :user_id => "1", :segmentQuery => segmentQuery } }
@@ -499,6 +507,26 @@ module ForestLiana
                 end
               end
 
+              context 'when received union segments NOT passing validation' do
+                let(:segmentQuery) { 'SELECT * FROM sellers/*MULTI-SEGMENTS-QUERIES-UNION*/ UNION SELECT column_name(s) FROM table1 UNION SELECT column_name(s) FROM table2' }
+                it 'should return false' do
+                  expect(subject.is_authorized?).to be false
+                end
+              end
+
+              context 'when received union segments passing validation' do
+                let(:segmentQuery) { 'SELECT * FROM sellers/*MULTI-SEGMENTS-QUERIES-UNION*/ UNION SELECT * FROM products' }
+                it 'should return true' do
+                  expect(subject.is_authorized?).to be true
+                end
+              end
+              context 'when received union segments with UNION inside passing validation' do
+                let(:segmentQuery) { 'SELECT COUNT(*) AS value FROM products/*MULTI-SEGMENTS-QUERIES-UNION*/ UNION SELECT column_name(s) FROM table1 UNION SELECT column_name(s) FROM table2' }
+                let(:segments_permissions) { ['SELECT COUNT(*) AS value FROM products;', 'SELECT column_name(s) FROM table1 UNION SELECT column_name(s) FROM table2;', 'SELECT * FROM products;', 'SELECT * FROM sellers;'] }
+                it 'should return true' do
+                  expect(subject.is_authorized?).to be true
+                end
+              end
             end
 
             context 'when user has not the required permission' do